+++ .gitignore
@@ -27,6 +27,7 @@ target/
 *.iws
 out/
 /prjProvartcc/.idea
+/out/
 
 # Eclipse
 .classpath
