--- prjProvartcc/src/main/resources/META-INF/persistence.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<persistence xmlns="http://xmlns.jcp.org/xml/ns/persistence"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/persistence http://xmlns.jcp.org/xml/ns/persistence/persistence_2_1.xsd"
-	version="2.1">
-	<persistence-unit name="jpa_prjProva" transaction-type="RESOURCE_LOCAL">
-		<provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-		
-         <class>model.Recurso</class>
-         <class>model.Opcao</class>
-         <class>model.Questao</class>
-		 <class>model.Disciplina</class>
-         <class>model.Topico</class>
-         <class>model.Tag</class>
-         
-		<properties>
-			<property name="javax.persistence.jdbc.url"
-				value="jdbc:mysql://localhost:3306/bdprova" />
-			<property name="javax.persistence.jdbc.driver" value="com.mysql.cj.jdbc.Driver"/>
-
-			<property name="javax.persistence.jdbc.user" value="root" />
-			<property name="javax.persistence.jdbc.password" value="password" />
-			<property name="hibernate.hbm2ddl.auto" value="update"/>
-			<property name="hibernate.dialect"
-				value="org.hibernate.dialect.MySQL5Dialect" />
-		</properties>
-	</persistence-unit>
-</persistence>
