--- prjProvartcc/src/model/ModelException.java
@@ -1,7 +1,0 @@
-package model;
-
-public class ModelException extends Exception {
-	public ModelException(String msgErro) {
-		super(msgErro);
-	}
-}
