--- prjProvartcc/src/model/NivelQuestao.java
@@ -1,5 +1,0 @@
-package model;
-
-public enum NivelQuestao {
-	FACIL, INTERMEDIARIA, DIFICIL;
-}
