--- prjProvartcc/src/model/TipoQuestao.java
@@ -1,5 +1,0 @@
-package model;
-
-public enum TipoQuestao {
-	RESPOSTA_UNICA, RESPOSTA_MULTIPLA, ASSERCAO_RAZAO;
-}
