+++ src/controller/CtrlPrograma.java
@@ -1,0 +1,36 @@
+package controller;
+
+import model.*;
+import model.dao.*;
+
+public class CtrlPrograma {
+
+	public static void main(String[] args) {
+
+		try {
+//			Disciplina d = new Disciplina("PROG1","Programacao de Computadores",3, "Aprender principíos de Programacao");
+//			DaoDisciplina daoDisciplina = new DaoDisciplina();
+//			daoDisciplina.salvar(d);
+
+//			Topico t = new Topico(1, "Herança", "Explicação sobre herança", d);
+//			DaoTopico dao = new DaoTopico();
+//			dao.salvar(t);
+
+			//Opcao o = new Opcao("A opção é a seguinte lalala", true);
+			//DaoOpcao dao = new DaoOpcao();
+			//dao.salvar(o);
+
+//			Questao q = new Questao("Instrução inicial", "Suporte", "Comando",
+//					NivelQuestao.DIFICIL, TipoQuestao.RESPOSTA_UNICA, true);
+//			DaoQuestao dao = new DaoQuestao();
+//			dao.salvar(q);
+
+			Tag t = new Tag("Teste", "Assunto do teste");
+			DaoTag dao = new DaoTag();
+			dao.salvar(t);
+
+		} catch (ModelException e) {
+			e.printStackTrace();
+		}
+	}
+}
