+++ src/model/ModelException.java
@@ -1,0 +1,7 @@
+package model;
+
+public class ModelException extends Exception {
+	public ModelException(String msgErro) {
+		super(msgErro);
+	}
+}
