+++ src/model/NivelQuestao.java
@@ -1,0 +1,5 @@
+package model;
+
+public enum NivelQuestao {
+	FACIL, INTERMEDIARIA, DIFICIL;
+}
